+++ CMakeLists.txt
@@ -24,8 +24,9 @@
 ################################################################################
 
 cmake_minimum_required ( VERSION 3.5.0 )
-
 project ("rocm-validation-suite" )
+enable_testing()
+include ( CTest )
 
 message(STATUS "CMAKE_BINARY_DIR = ${CMAKE_BINARY_DIR}")
 message(STATUS "CMAKE_CURRENT_SOURCE_DIR = ${CMAKE_CURRENT_SOURCE_DIR}")
@@ -164,5 +165,10 @@ add_custom_target(userguide ALL
 
 install(FILES "${ROCM_SMI_LIB_DIR}/lib${ROCM_SMI_LIB}.so" DESTINATION ${CMAKE_PACKAGING_INSTALL_PREFIX}/rvs/lib COMPONENT rvsmodule)
 install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/rvsso.conf" DESTINATION /etc/ld.so.conf.d COMPONENT rvsmodule)
+
+# TEST SECTION
+
+add_test(NAME test1 COMMAND rvs)
+
 
 include (CPack)
