+++ rvs/conf/pebb_test6.conf
@@ -1,0 +1,26 @@
+# REBB test #5
+# testing conditions:
+# 1. all AMD compatible GPUs
+# 2. all types of devices
+# 3. host to device
+# 4. parallel transfers
+#
+# Run test with:
+#   cd bin
+#   sudo ./rvs -c conf/pebb_test6.conf -d 3
+#
+# Expected result:
+#   Test shows bandwidth from host to device and device to host based on devi
+# deviceid and devices
+# FALSE otherwise
+
+actions:
+- name: h2d-d2h
+  device: all
+  duration: 5000
+  module: pebb
+  log_interval: 1000
+  device_to_host: true
+  host_to_device: false
+  block_size: 512000000
+  count: 10000
